Squash of Srdan's commits
Squash starting from 094487aa34ddfa750d408cf64fbb509e8aeb56a9.

Then:
- Whitespace-only changes have been reverted.
- EOL whitespace trimmed, EOF newline fixed.
- Spelling.
- Suspiciously empty source files removed.
- Remove a few renames that were covered by the foils refactor (592ef529817ea710948e632b40e9fbde8c655c48)
- #includes have been transformed to use relative header paths (479079ec0c5b478239dcaf5f57935c6af763aa7f)
- Remove suspiciously empty source files
- Convert to Meson build system (382bb3e0fc7d9a527e9685df09222d427dd3718a)
- Remove a rename that was covered by the level0 foil refactor (565d7281e54a3c7fdf9eae03131d411aa9843862)
- Update copyright notices to 2024. (5cdede372c887f824b26bad3cb70071dfdd3c68d)
- Update includes for libmonitor integration (9b9f07bc96907a11d967bbc203f76c56fce326da)
- Add #defines for preprocessor cleanup (31e37be1b8821eaa1c70d8e9707d0927ebf40c8c)

+++ src/tool/hpcrun/gpu/gpu-monitoring-thread-api.h
@@ -44,6 +44,12 @@
 #ifndef gpu_monitoring_thread_api_h
 #define gpu_monitoring_thread_api_h
 
+//******************************************************************************
+// system includes
+//******************************************************************************
+
+#include <stdint.h>
+
 
 
 //******************************************************************************
@@ -60,9 +66,8 @@ gpu_monitoring_thread_activities_ready
 void
 gpu_monitoring_thread_activities_ready_with_idx
 (
- int idx
+ uint64_t idx
 );
-
 
 
 #endif
